Refactored Header components into folder
