+++ src/web/components/Header/ProfilePicture.js
