+++ src/web/components/Templates/TemplateHeader.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Container } from 'reactstrap';
 
 import Member from '../../../containers/Member';
-import Header from '../Header';
+import Header from '../Header/Header';
 import Footer from '../Footer';
 
 const Template = ({ children }) => (
